cpp(BREAKING)!: support both global-level and messager-level load options (#117)
* refactor: support both global-level and messager-level load options

* fix: test

* refactor: cpp loader codegen

* fix: cpp loader codegen

* feat: support at least c++17

* fix: gen.bat

* fix: update CMakeLists and fix compile warnings

* fix: some compile warnings in MSVC

* dev(linux): support at least c++17

---------

+++ cmd/protoc-gen-cpp-tableau-loader/embed/scheduler.pc.cc
@@ -26,7 +26,7 @@ int Scheduler::LoopOnce() {
   for (auto&& job : jobs) {
     job();
   }
-  count += jobs.size();
+  count += static_cast<int>(jobs.size());
   return count;
 }
